requirements.txt
numpy
pandas
plotly
dash
dash_bootstrap_components
sqlalchemy
psycopg2-binary
python-dotenv
Pillow
